moved tpuartserial and ncn5120 to the same base class - only small things are different
