+++ src/backend/ncn5120.h
@@ -21,70 +21,25 @@
 #define SNCN5120_SERIAL_H
 #include <termios.h>
 #include "lowlatency.h"
-#include "layer2.h"
+#include "tpuartserialbase.h"
 
 /** TPUART user mode driver */
-class NCN5120SerialLayer2Driver : public Layer2Interface, private Thread
+class NCN5120SerialLayer2Driver : public TPUARTBASESerialLayer2Driver
 {
-  /** old serial config */
-  low_latency_save sold;
-  /** old termios state */
-  struct termios old;
-  /** file descriptor */
-  int fd;
-  /** debug output */
-  Trace *t;
-  /** default EIB address */
-  eibaddr_t addr;
-  /** state */
-  int mode;
-  /** vbusmonitor mode */
-  int vmode;
-  /** semaphore for inqueue */
-  pth_sem_t in_signal;
-  /** semaphore for outqueue */
-  pth_sem_t out_signal;
-  /** input queue */
-    Queue < LPDU * >inqueue;
-    /** output queue */
-    Queue < LPDU * >outqueue;
-    /** event to wait for outqueue */
-  pth_event_t getwait;
-  /** my individual addresses */
-    Array < eibaddr_t > indaddr;
-    /** my group addresses */
-    Array < eibaddr_t > groupaddr;
-  bool ackallgroup;
-  bool ackallindividual;
-  bool dischreset;
-
-    /** process a recevied frame */
-  void RecvLPDU (const uchar * data, int len);
-  void Run (pth_sem_t * stop);
 public:
     NCN5120SerialLayer2Driver (const char *dev, eibaddr_t addr, int flags,
 			      Trace * tr);
    ~NCN5120SerialLayer2Driver ();
-  bool init ();
 
-  void Send_L_Data (LPDU * l);
-  LPDU *Get_L_Data (pth_event_t stop);
+protected:
+  struct termios getTermiosSettings();
+  int getBaudRate();
 
-  bool addAddress (eibaddr_t addr);
-  bool addGroupAddress (eibaddr_t addr);
-  bool removeAddress (eibaddr_t addr);
-  bool removeGroupAddress (eibaddr_t addr);
+  void afterLPDUReceived();
+  bool beforeLoop(CArray &in);
 
-  bool enterBusmonitor ();
-  bool leaveBusmonitor ();
-  bool openVBusmonitor ();
-  bool closeVBusmonitor ();
-
-  bool Open ();
-  bool Close ();
-  eibaddr_t getDefaultAddr ();
-  bool Connection_Lost ();
-  bool Send_Queue_Empty ();
+private:
+  bool m_rmn;
 };
 
 #endif
